+++ include/sealib/dictionary/sharedrankselect.h
@@ -29,14 +29,14 @@ class SharedRankSelect {
      * @param k idx
      * @return k-th set bit
      */
-    uint64_t select(uint64_t k) const;
+    uint select(uint k) const;
 
     /**
      * Rank of the k-th idx
      * @param k idx
      * @return rank of k-th idx
      */
-    uint64_t rank(uint64_t k) const;
+    uint rank(uint k) const;
 
     ~SharedRankSelect();
 };
